refactor: Io and AddressUtil strings to common
