+++ wallet/proguard.cfg
@@ -94,7 +94,7 @@
 -dontnote ch.qos.logback.core.rolling.helper.FileStoreUtil
 
 # Bitcoin Wallet
--dontnote de.schildbach.wallet.util.Io
+-dontnote org.dash.wallet.common.util.Io
 
 -keepattributes InnerClasses
